+++ src/nn/evolution/tests/mod.rs
@@ -12,3 +12,4 @@ mod node_gene;
 mod proxy;
 mod selection;
 mod task;
+mod asha;